requirements.txt
openai==2.24.0
python-dotenv==1.2.2
seleniumbase==4.47.1
beautifulsoup4==4.14.3
